Split home-content.tsx into focused section modules
Extract 5 components from the 957-line monolith into app/sections/:
- mock-data.ts — shared USERS, COMMANDS, TAGS data
- section-heading.tsx — clickable anchor heading with hash link
- demo-section.tsx — hero demo with PromptArea + ActionBar
- features-grid.tsx — feature card grid
- inspector-section.tsx — interactive playground with event log
- dark-theme-preview.tsx — light/dark side-by-side comparison

home-content.tsx is now a thin composition file (~250 lines).
