+++ app/sections/mock-data.ts
@@ -1,0 +1,35 @@
+export const USERS = [
+  { value: 'alice', label: 'Alice', description: 'Engineering' },
+  { value: 'bob', label: 'Bob', description: 'Design' },
+  { value: 'charlie', label: 'Charlie', description: 'Product' },
+  { value: 'diana', label: 'Diana', description: 'Marketing' },
+  { value: 'eve', label: 'Eve', description: 'Sales' },
+]
+
+export const COMMANDS = [
+  {
+    value: 'summarize',
+    label: 'summarize',
+    description: 'Summarize the conversation',
+  },
+  {
+    value: 'translate',
+    label: 'translate',
+    description: 'Translate to another language',
+  },
+  {
+    value: 'improve',
+    label: 'improve',
+    description: 'Improve writing quality',
+  },
+  { value: 'explain', label: 'explain', description: 'Explain a concept' },
+  { value: 'code', label: 'code', description: 'Generate code snippet' },
+]
+
+export const TAGS = [
+  { value: 'bug', label: 'bug' },
+  { value: 'feature', label: 'feature' },
+  { value: 'docs', label: 'docs' },
+  { value: 'urgent', label: 'urgent' },
+  { value: 'question', label: 'question' },
+]
